Tipologia

+++ src/main/java/com/generation/loja_de_games/repository/ProdutoRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,6 +14,6 @@ public interface ProdutoRepository extends JpaRepository<Produtos, Long> {
 
     public List<Produtos> findAllByNomeContainingIgnoreCase(@Param("nome") String nome);
 
-    public List<Produtos> findAllByPrecoGreaterThanOrderBOrderByPreco(float preco);
-    public List<Produtos> findAllByPrecoLessThanOrderByPrecoDesc(float preco);
+    public List<Produtos> findAllByPrecoGreaterThanOrderBOrderByPreco(BigDecimal preco);
+    public List<Produtos> findAllByPrecoLessThanOrderByPrecoDesc(BigDecimal preco);
 }
